Save before implementing saving of offers

+++ PcBuilderApi/Models/CpuCooler.cs
@@ -44,10 +44,10 @@ namespace PcBuilderApi.Models
         public double? NoiseLevelDb { get; set; }
         public int? Voltage { get; set; }
         public int? Lifespan { get; set; }
-        public int? Length { get; set; }
-        public int? Width { get; set; }
-        public int? Height { get; set; }
-        public int? Weight { get; set; }
+        public double? Length { get; set; }
+        public double? Width { get; set; }
+        public double? Height { get; set; }
+        public double? Weight { get; set; }
         public int? Wattage { get; set; }
 
         [Url]
